+++ common/Extend/UIImage+Ex.m
@@ -71,12 +71,12 @@
     }
 }
 
-- (UIImage *)aspectToScale:(CGFloat)scaleWidth {
-    CGFloat oldWidth = self.size.width;
-    CGFloat scaleFactor = scaleWidth / oldWidth;
+- (UIImage *)aspectToScale:(float)scaleWidth {
+    float oldWidth = self.size.width;
+    float scaleFactor = scaleWidth / oldWidth;
     
-    CGFloat newWidth = oldWidth * scaleFactor;
-    CGFloat newHeight = self.size.height * scaleFactor;
+    float newWidth = oldWidth * scaleFactor;
+    float newHeight = self.size.height * scaleFactor;
     
     UIGraphicsBeginImageContext(CGSizeMake(newWidth, newHeight));
     [self drawInRect:CGRectMake(0, 0, newWidth, newHeight)];
